+++ src/has-magic.ts
@@ -2,7 +2,15 @@ import { Glob, GlobOptions } from './glob.js'
 
 /**
  * Return true if the patterns provided contain any magic
- * glob characters.
+ * glob characters, given the options provided.
+ *
+ * Brace expansion is not considered "magic" unless the `magicalBraces` option
+ * is set, as brace expansion just turns one string into an array of strings.
+ * So a pattern like `'x{a,b}y'` would return `false`, because `'xay'` and
+ * `'xby'` both do not contain any magic glob characters, and it's treated the
+ * same as if you had called it on `['xay', 'xby']`. When
+ * `magicalBraces:true` is in the options, brace expansion _is_ treated as a
+ * pattern having magic.
  */
 export const hasMagic = (
   pattern: string | string[],
@@ -11,8 +19,10 @@ export const hasMagic = (
   if (!Array.isArray(pattern)) {
     pattern = [pattern]
   }
+  const length = pattern.length
   const g = new Glob(pattern, options)
-  return g.patterns.length === 0
-    ? false
-    : g.patterns.some(p => p.hasMagic())
+  if (g.magicalBraces && g.patterns.length !== length) {
+    return true
+  }
+  return g.patterns.some(p => p.hasMagic())
 }